Move view models to com.jerboa.model
