+++ app/src/main/java/com/jerboa/model/CreatePostViewModel.kt
@@ -1,4 +1,4 @@
-package com.jerboa.ui.components.post.create
+package com.jerboa.model
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
