+++ app/src/main/java/com/jerboa/model/PostEditViewModel.kt
@@ -1,4 +1,4 @@
-package com.jerboa.ui.components.post.edit
+package com.jerboa.model
 
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
