+++ app/src/main/java/com/jerboa/ui/components/community/CommunityActivity.kt
@@ -51,6 +51,7 @@ import com.jerboa.db.AccountViewModel
 import com.jerboa.db.AppSettingsViewModel
 import com.jerboa.isLoading
 import com.jerboa.isRefreshing
+import com.jerboa.model.CommunityViewModel
 import com.jerboa.newVote
 import com.jerboa.scrollToTop
 import com.jerboa.shareLink
@@ -70,7 +71,7 @@ import com.jerboa.ui.components.common.toPost
 import com.jerboa.ui.components.common.toPostEdit
 import com.jerboa.ui.components.common.toPostReport
 import com.jerboa.ui.components.common.toProfile
-import com.jerboa.ui.components.home.SiteViewModel
+import com.jerboa.model.SiteViewModel
 import com.jerboa.ui.components.post.PostListings
 import com.jerboa.ui.components.post.edit.PostEditReturn
 import kotlinx.collections.immutable.toImmutableList
