+++ app/src/main/java/com/jerboa/ui/components/home/sidebar/SiteSidebarActivity.kt
@@ -10,7 +10,7 @@ import com.jerboa.ui.components.common.ApiEmptyText
 import com.jerboa.ui.components.common.ApiErrorText
 import com.jerboa.ui.components.common.LoadingBar
 import com.jerboa.ui.components.common.SimpleTopAppBar
-import com.jerboa.ui.components.home.SiteViewModel
+import com.jerboa.model.SiteViewModel
 
 @OptIn(ExperimentalMaterial3Api::class)
 @Composable
