+++ app/src/main/java/com/jerboa/ui/components/post/PostActivity.kt
@@ -107,7 +107,7 @@ import com.jerboa.ui.components.common.toPost
 import com.jerboa.ui.components.common.toPostEdit
 import com.jerboa.ui.components.common.toPostReport
 import com.jerboa.ui.components.common.toProfile
-import com.jerboa.ui.components.home.SiteViewModel
+import com.jerboa.model.SiteViewModel
 import com.jerboa.ui.components.post.edit.PostEditReturn
 
 @Composable
