+++ app/src/main/java/com/jerboa/ui/components/privatemessage/PrivateMessageReplyActivity.kt
@@ -19,11 +19,12 @@ import androidx.navigation.NavController
 import com.jerboa.api.ApiState
 import com.jerboa.datatypes.types.PrivateMessageView
 import com.jerboa.db.AccountViewModel
+import com.jerboa.model.PrivateMessageReplyViewModel
 import com.jerboa.ui.components.common.InitializeRoute
 import com.jerboa.ui.components.common.LoadingBar
 import com.jerboa.ui.components.common.getCurrentAccount
 import com.jerboa.ui.components.common.toProfile
-import com.jerboa.ui.components.home.SiteViewModel
+import com.jerboa.model.SiteViewModel
 
 @Composable
 fun PrivateMessageReplyActivity(
