+++ app/src/main/java/com/jerboa/ui/components/settings/account/AccountSettings.kt
@@ -30,7 +30,7 @@ import com.jerboa.datatypes.types.SortType
 import com.jerboa.db.Account
 import com.jerboa.imageInputStreamFromUri
 import com.jerboa.ui.components.common.*
-import com.jerboa.ui.components.home.SiteViewModel
+import com.jerboa.model.SiteViewModel
 import com.jerboa.ui.theme.*
 import kotlinx.coroutines.launch
